+++ src/cores/elastic/hangul-service.ts
