+++ src/cores/elastic/index.ts
@@ -1,0 +1,12 @@
+/**
+ * `cores/elastic/index.tx`
+ * - default index to export all for elastic-search
+ *
+ * @author      Steve Jung <steve@lemoncloud.io>
+ * @date        2022-05-13 initial file to group
+ *
+ * @copyright (C) lemoncloud.io 2022 - All Rights Reserved.
+ */
+export * from './elastic6-service';
+export * from './elastic6-query-service';
+export * from './hangul-service';
