+++ src/cores/index.ts
@@ -15,16 +15,15 @@ export * from './aws/';
 export * from './config/';
 export * from './lambda/';
 export * from './protocol/';
+export * from './dynamo/';
+export * from './elastic/';
 
 //! export services
 export * from './api-service';
 export * from './redis-storage-service';
-export * from './elastic6-service';
-export * from './elastic6-query-service';
 export * from './storage-service';
 export * from './proxy-storage-service';
 export * from './cache-service';
-export * from './hangul-service';
 export * from './model-manager';
 
 //! import default with named.
