+++ src/cores/lambda/lambda-dynamo-stream-handler.ts
@@ -14,8 +14,8 @@ import { DynamoDBRecord } from 'aws-lambda';
 import { LambdaHandler, DynamoStreamHandler, LambdaSubHandler, buildReportError } from './lambda-handler';
 import { NextHandler } from '../core-types';
 import { toJavascript } from '../../lib/dynamodb-value';
-import { Elastic6Service, Elastic6Item } from '../elastic6-service';
-import { DynamoOption } from '../dynamo-service';
+import { Elastic6Service, Elastic6Item } from '../elastic/';
+import { DynamoOption } from '../dynamo/';
 const NS = $U.NS('HDBS', 'green'); // NAMESPACE TO BE PRINTED.
 
 export type DynamoStreamEvent = DynamoDBRecord['eventName'];
